Convert list of links into hash of links
Each resource has an array of hashes of links, where each links hash
has a 'rel' and an 'href'. Convert this into a single hash where the
rels (converted to symbols) are the keys and the hrefs are the
values. NB: This assumes that there is only one link of each rel
type in each resource!

Implement Resource#link to lookup a link by its rel.